requirements.txt
alembic==1.7.5
antlr4-python3-runtime==4.8
attrs==21.4.0
autopage==0.4.0
Brotli==1.0.9
click==8.0.3
cliff==3.10.0
cmaes==0.8.2
cmd2==2.3.3
colorlog==6.6.0
cycler==0.11.0
dash==2.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.2
Flask-Compress==1.10.1
fonttools==4.28.5
greenlet==1.1.2
h5py==3.6.0
hydra-core==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
Mako==1.1.6
mariadb==1.0.9
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.0
omegaconf==2.1.1
optuna==2.10.0
packaging==21.3
pandas==1.4.0
pbr==5.8.0
Pillow==9.0.0
plotly==5.5.0
prettytable==3.0.0
pyarrow==6.0.1
PyMySQL==1.0.2
pyparsing==3.0.6
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.29
stevedore==3.5.0
tenacity==8.0.1
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
typing_extensions==4.0.1
wcwidth==0.2.5
Werkzeug==2.0.2
